Add About button
The about button pop ups a window telling the user what version of AToMPM they are running, as well as the website and
documentation location.

Also, the about menu contacts the GitHub release site to determine the newest version and its release date. If no
connection can be made, this information is not presented.