Add type matching to semantic check in match algorithm.
When checking if two nodes are semantically compatible in the match algo.,
check if the types match or if the src node is a sub-type of the pattern node.

This was causing problems with pivots. Specifically, if a super class
was in the LHS of a rule, and a sub-type in the NAC, then the rule was
prevented from firing in too many cases.