Be more precise about which changes to a model mark it as unsaved.
SYSOUT messages which are used to print to the console were
marking the model as unsaved. This is used, for example,
for status messages from the transformation server.

Now, the list of changes which mark the model as unsaved is
explicitly recorded.

Note: The backdoor API is operated as a SYSOUT change, so this
is specially handled.