Raise Exception for default code in rules
If default code (looking like [PYTHON] Example: result = True Example: result = getAttr()...)
is left in a rule, then the rule will fail to match (or not match).

This commit adds a check during Himesis graph creation that the node's value is not this default code.
If so, an exception is thrown so that the transformation can't be loaded.

Future work should address the issue of association values not automatically changing when
they are contained in a LHS/NAC/RHS.