Mark all changes as making model unsaved
There is a bug where making changes to an element does not
report these changes in the changelog steps. Such as changing
the condition code in a rule LHS.

Therefore the earlier commits are incorrect. This commit
partially reverts the earlier one by adding in an express
call to note the model as unsaved on any changelog modification.