Add createSession to session_manager.
The client now uses socket to ask the session manager for a new session.

This allows the session manager to map the client id to the csw id,
and the socket id to the csw id for returning changelogs.

Joining a session for collaboration is not implemented.

This commite also refactors out the creation of workers.