Session manager handles both workers when creating a session
Before this, the client was directing the CSWorker to subscribe to the ASWorker.
Now, the SessionManager directly creates both workers.

The CSWorker is then directed to subscribe. This does involve a Socket call
from the CSWorker to the SessionManager, but this is because the SessionManager
needs a Socket connection from the CSWorker to know the CSWorker's port.