Add collaboration button and dialog (#160)
Adds a button to the main menu which opens a collaboration dialog, and removes the collaboration links in the bottom-right of AToMPM.

This dialog explains the screenshare/modelshare collaboration options, and offers buttons to copy the collab urls to the clipboard. This is a much easier workflow than the old method and actually provides some explanation to the user. The link information is also generated by the server-side (in session_manager.js).